Skip CallSequencingManagedHoldRestrictionTest when telecom not enabled
We are not properly skipping the CallSequencingManagedHoldRestrictionTest CUJ test on devices that don't support Telecom like TV. Upon closer inspection, it looks like we're properly skipping setUp but the test is still run even though we verify if we should test Telecom via assumeTrue(). Add an explicit check to the test methods so that these are properly skipped.

Bug: 409956980
Flag: EXEMPT test fix
Test: atest CallSequencingManagedHoldRestrictionTest
Change-Id: If7262d71adec02694b263bc5db767a7d3709b47a