Update NumberBlockingTest to fix issues.
1. The test was not running any more due to incorrect feature flag tests.
2. When it was running the secondary users check was not working at all
due to the fact that the underlying
"testIncomingCallFromBlockedNumberIsRejected" test doesn't register a
CAPABILITY_MULTI_USER phone account, so blocking across users is not
expected to actually work.  This part of the test will need to be reworked
in the future, so disabling it for now.
3. Updated to use "getMainUserId" since "getPrimaryUserId" is deprecated
and this is the suggested replacement.

Bug: 408967027
Flag: EXEMPT test only fix.
Test: Re-run test and verify successful completion.
Change-Id: I3e579393cf5814dfbd0e6fa6219c8859729e236b