Improve DreamServiceTest stability.
A subset of the DreamServiceTest suite is flaky due to inconsistency
receiving focus change signal from the dream window. This changelist
addresses this issue by changing the tests to instead verify the
following:
- input focus gained: We now listen to the content view where we can
add the listener at construction before it is set as the window
content.
- input focus lost: In the test cases of interest, we were tracking
this signal to know that we had moved focus to the keyguard. We can
directly assess this state as well by making sure the keyguard is
showing and occluded.

Test: atest DreamServiceTest
Fixes: 365664732
Flag: EXEMPT bugfix
Change-Id: I8124beedf82415c6bcddd0fbe50521b9e66f6289