Wait for stable window geometry in SimultaneousTouchAndStylusTest
We are seeing occasional failures in SimultaneousTouchAndStylusTest.
In the failing cases, the activity is getting the touch instead of one
of the two windows that are overlaid. We are already waiting for the
windows to be visible, but perhaps their dimensions are not finalized
yet when the touch occurs.

Try to fix it in this CL by waiting for stable window geometry.

Bug: 413292550
Test: SimultaneousTouchAndStylusTest
Flag: TEST_ONLY
Change-Id: I08950d1b04bef7f2a5f6e2bb3e636ffe708e7f70