Increase threshold when checking the color
The configuration for the color is ARGB_8888. So the actual RGB comparsion is 0xff0000 (pure red 100 R, 0 G, 0 B) vs
ee0e0e (93 R, 5 G, 5 B).
The threshold is 30,while the actual error is 45.

Similar thing happens in all 4 corners. The colors are close enough,
but the error are slightly bigger. For instance #0000FF vs #1010ee
(100 blue vs 93 blue)

The test case is only for testing if YUB->RGB are handled and
decoder accuracy is covered in other test cases.

It can pass after increase the threshold to 50.

Bug: 403329421
Test: atest android.view.cts.PixelCopyTest#testVideoProducer
Flag: EXEMPT bug fix

Change-Id: Icdc6199daffac7403cfc09569c78f64371a60eab