Add CTS for new RenderParams form options
This covers RENDER_FORM_CONTENT_ENABLED and RENDER_FORM_CONTENT_DISABLED
for both PdfRenderer and PdfRendererPreV

RENDER_FORM_CONTENT_DEFAULT is not covered by CTS, but 1) it's the
current behavior and 2) it is covered by existing unit tests, see
PdfCompatChangesTest in MediaProvider. That test relies on the compat
changes framework, and we moved it off of CTS after significant issues
(see b/352793202). It may eventually be moved back to CTS after a
rewrite, but that's orthogonal to these new APIs.

Bug: 410009817
Bug: 413456220
Test: atest CtsPdfModuleTestCases
Flag: android.graphics.pdf.flags.enable_render_params_form_options
Change-Id: I152d6484fc2f78c6c2731237ccf9c329fb11789b