Fix telecom audio focus lock CUJ test
The test should verify that audio focus is held by Telecom while the
call is still active. Currently, the test will attempt to set the call
disconnected and then verify that audio focus lock is retained. This is
not failing on Pixel devices due to the timing of when audio focus is
abandoned as a result of the disconnect.

Bug: 413283599
Flag: EXEMPT test fix
Test: atest SingleCallingTest
Change-Id: I9072cc5655c04057c868e6d8428f1b0cb1241783