Improve flakiness of self-managed CS related CTS tests.
HandoverTest#testIncomingHandoverRequestValid is exceeding the flake
threshold for platinum.  I did some investigation in the logs and it
looks like the issue is that although we wait for the InCallService to
report the call before continuing, we DO NOT wait until the
ConnectionService#onCreateconnectionComplete fires.  In some cases it
appears the InCallService hears of the call just before it gets fully
added to Telecom, and hence some of the immediate actions we try to perform
in the tests don't get through to Telecom.

Fixes: 413087434
Flag: EXEMPT test only fix.
Test: THIS is a test fix.
Change-Id: Ie2a8fe95f28c015e1989bed4f3cb3059af97a181