Validate emergency call is in dialing state before activating it.
During the emergency call sequencing test, the call may still be connecting when we try to verify that the call has gone active so verifying that it's in dialing first should resolve the flakiness we observe with these failures. Do the same with the normal call as well.

Bug: 413043660
Flag: EXEMPT test fix
Test: atest CallSequencingBasicEmergencyCallTest
Change-Id: If9b1153c0d9cdbe57b09a2bf8f0b107c3415a6ff