Improve stability of ScreeningServiceWithContactsPermissionTest CTS.
Flakes are seen with the tests in this particular class.  It appears in
some cases that the tests are acting like there is a contact present when
there is supposed to be none.
The CTS tests were always inserting 6505551212 as a test contact into the
contacts database and using that to verify; there are other telecom CTS
which use that same number; any flakiness in those tests resulting in a
failure to clean things up could cause a problem here.  Additionally any
pre-existing contacts on a device that were lingering may cause a problem
as well.

This change instead uses a random +1 prefix phone number and verifies
that it does not already exist in contacts (if it does a new one is
chosen).  The contact entry is deleted at the end of the test, but this
will ensure no other tests using the same number can interfer.

Test: This is a test improvements
Flag: EXEMPT test only fix.
Fixes: 414586335
Fixes: 414038451
Fixes: 413152990
Change-Id: Ia8cbf59371a891acb75e5004ea480faf2b3d8d32