remove executable_method_file_offsets flag check in tests
This change removes the executable_method_flag_offses flags from the
test. The reason is CTS should not depends on exported flags since in
the release build of CTS, the exported flags accessor will still read
the flag value from device. However for a release build of CTS, the CTS
should check the expected behavior set in the tests but not reading the
flag from the devices.

These two flags are enabled in B release, so it is safe to switch the
flag check to sdk level check.

Test: CtsDynamicInstrumentationManagerTest
Bug: 303773055
Flag: EXEMPT remove flag
Change-Id: I5376eaef5358149b5c70eb7b58a731c96b7bb085