[cts] Fix the IncFS block size calculation on 16k devices
Before 16kb pages the incfs block size was equal to the
page size everywhere on the device, so the test used the
filesystem's page size assuming it's the same. Now it's
wrong, and makes the test significantly overestimate the
expected preallocation. This CL just hardcodes the constant.

Bug: 414454999
Test: atest PackageManagerShellCommandIncrementalTest
Flag: EXEMPT test fix
Change-Id: Ic1c054b9150ee0944c957a3a952310b9dccd6eaa