Update tests which depended on the fgs boot completed flag.
The flag is now being removed so we can remove tests which were meant to
run when the flag was disabled and remove the enabled flag since the
flag is always enabled.

Also, remove the annotations in ServiceTest since the test are unrelated
to the flag.

Fixes: 409605469
Test: atest BootCompletedFgsStartTest
Test: atest ServiceTest
Flag: EXEMPT removing com.android.server.am.fgs_boot_compelted
Change-Id: Iffa0e4b392afd429f3e3142bccec5451a95482db