remove platform_apis: true settings from cts test
This change removes platform_apis: true from this cts test, and set the
sdk_version to test_current. CTS should not depend on platform_apis,
since platform_apis: true means the test can use the hidden apis in the
framework. CTS should test the APIs, and compatibility behaviors of the
devices not but rely on the implementation detail of the devices.

Test: atest CarSecurityPermissionTest
Bug: 303773055
Flag: EXEMPT build file change
Change-Id: Id002711772370b8ba6c9d1913065b329a5eab385