[CTS] Fix flaky testActivityFinishAndLaunchNewActivity
The order of the activity.finish() call and the corresponding
TFO callback is undefined. If TFO callback comes first, the old
split container is removed and the new activity will form a split
using a new split container. This is still a correct ending state
but it results in an additional SplitInfo callback, which causes
test flakiness. This change ignores the additional callback in
the test case.

Bug: 402374035
Test: atest ActivityEmbeddingLaunchTests
Flag: com.android.window.flags.activity_embedding_delay_task_fragment_finish_for_activity_launch
Change-Id: Ib568013d2346a4f557993f0ccbcad22dbff4cad5