WebView: check for Activity focus and screen unlocked
This checks if the screen is off (or if the CTS Activity is not in
focus) and throws a CtsRequirementException.

This is hopefully redundant with ag/33244601 which is designed to force
the screen on and unlocked, however I'm landing this check as well as
defense in depth.

Bug: 415120568
Flag: TEST_ONLY
Test: atest --instant CtsWebkitTestCases
Test: atest CtsWebkitTestCases (with screen on)
Test: atest CtsSdkSandboxWebkitTestCases (with screen on)
Test: atest CtsHostsideWebViewTests (with screen on)
Test: try screen off (PowerManager catches this)
Test: try screen on but locked (KeyguardManager catches this)
Test: try screen on & unlocked but notification shade is down (polling for focus catches this)
Test: try screen on & unlocked & nothing stealing focus (test pass)
Change-Id: Ie2249af8f0cc1801837e68f8de4a38efabee33a6