Adding CUJ test coverage for DND behavior for VoIP calls.
Updated RingerTest to verify audio focus behavior for both transactional
and self managed CS Voip calls.  We now verify for incoming calls that we
do not get ringing audio focus when a call is suppressed by DND.

The changes to the transactional and smcs voip app are to expose
a way to send a connection event.  This is used so that we have a way of
sending _something_ from the app through to the InCallService via
Telecom.  This is done as a way to ensure that the audio mode change has
fully completed before we continue with verification.  In testing I found
that the ICS was updated quickly while the audio mode change lagged
slightly behind.  Adding the pass through of the connection event
ensures that the previous operation has finished and we can verify the
mode stuff as needed.

Bug: 415288638
Test: This is test!
Flag: com.android.server.telecom.flags.voip_dnd_focus
Change-Id: I3fd534fea47cdcb0ec6e5532545c7a5b4a30b4c3