Relax requirements of BatteryHealthTest#testBatterySerialNumber_dataValid
Testing that a battery serial number is at least 6 characters assumes
that all batteries actually have serial numbers that are at least that
long, which is false. For example, the marasov device with serial number
R8NXCV05W978323 has a battery with serial number 0x139C, which reads as
the string "139C".

Instead of assuming any particular length of string, simply require that
the serial number string is not empty because any other assumption may not
always be true.

Bug: 416256111
Flag: TEST_ONLY
Change-Id: I5163bad0f65a5e5873ae0e7c5c7896dd996d1160