CtsVerifier USB Companion app: Fix a bug in finding accessory interface
There is a bug in finding accessory interface in the app. The original
code tried to find the interface based on the number of device
configurations. However, getInterface() method of UsbDevice returns the
interface based on the index of all interfaces in a configuration
instead of the configuration. This CL is to fix the bug by correcting
using "getInterfaceCount()".

TEST: Run CTS verifier and pass the tests
Flag: TEST_ONLY
FIXES: 417099369
Change-Id: I3f385827d1c6cbcf5279d18a44d4a02c5e9d3cde