Hide IME before exiting split screen activity
In most cases, when exiting split screen to a single activity that does
not handle configuration changes, the activity will get re-created.
However, based on the device and display configuration, exiting split
screen would lead to a configuration change that is not significat
enough to be reported, which won't re-create the activity.

To make the test device agnostic, this will hide the IME before trying
to exit split screen, leaving it in a consistent state.

Flag: EXEMPT testfix
Bug: 413131452
Test: atest KeyboardVisibilityControlTest#testIMEVisibleInSplitScreenWithWindowInsetsApi
Change-Id: Ib37374f3770b5192efc8eccd535f5a28ef2feda5