Fix test failure in testModeSettingRestrictedToShellForDeviceProvisioningApp
The app op mode of OPSTR_RESERVED_FOR_TESTING for the device provisioning app is not always ALLOWED
as we've seen from test runs in the bug. Since the test is asserting we
canot change the mode, it doesn't matter what the original mode is.
Hence, change the test to be more tolerable of the original mode.

Bug: 412582350
Test: android.app.appops.cts.AppOpsTest#testModeSettingRestrictedToShellForDeviceProvisioningApp
Flag: EXEMPT TEST_ONLY
Change-Id: Ie2cb9e7115ac51307ec8e3a78e1d2083a38d0de1