CTSV Data Paths fix short analysis times, Timer.schedule
Tests were started at a fixed rate. But they can take a variable time to
open and close streams. So the time left for analysis was variable
and could be too short to get a valid measurement.
The new code use Timer.schedule() to always analyze for enough time.

Bug: 417567867
Test: CTS Verifier
Test: run all Audio Data Path tests
Flag: TEST_ONLY
Change-Id: I128e71d95f3751bf7a31c4977e9332cc7091c0ba