WebView: add custom TargetPreparer
No change to logic. This adds a custom TargetPreparer for WebView CTS
tests which runs during the test setup in order to fetch debugging info
about the WebView service. This runs 'adb shell dumpsys webviewupdate',
captures the output in a text file, and logs this file as a tradefed
artifact which will be visible on test infra.

The main goal of this change is to log the WebView version in a single
location which is easily discoverable.

The log file will follow a pattern like:
"log/invocation_*/inv_*/webview-version_*.txt".

This TestPreparer is integrated into CtsWebkitTestCases,
CtsSdkSandboxWebkitTestCases, and CtsHostsideWebViewTests.

Fixes: 417818812
Test: atest CtsWebkitTestCases
Test: atest CtsSdkSandboxWebkitTestCases
Test: atest CtsHostsideWebViewTests
Flag: TEST_ONLY
Change-Id: I3cbb5faf2a0b03fadcf6100dc6b493192cbef44e