Skip moveToBackground on watches.
If SysUI has the QSS or navigation bar with persistent flag, we can not
ensure that android.app.stubs.shared is on the top. Due to this reason,
a.moveToBacground failed if the SysUI has the persistent flag.

The activity will not be launching properly if the search service for
the intent action (search) called in the prior conditional statement
("android.software.car.splitscreen_multitasking") does not exist on the
target device. To avoid this, the main activity of the CTS test app is
used for testing.

Bug: 413155192 Testing:
- tested locally on wear.
- atest CtsAppServiceTestCases:android.app.cts.service.ServiceTest#testActivityServiceBindingLru (cuttlefish - mobile, auto, wear)
Flag: TEST_ONLY

Change-Id: Ibee86fdf0bae1fcb5b550a7016f9df8c059d4c45