Move locktask into separate pkg
Currently the activity that enters lock task mode is located in the
instrumented package. As a result, when host-side starts a device-side
test to ensure the lock task is running after reboot, that activity
gets killed and has to be restarted by the test, which defeats its
purpose - it can pass even if the activity wasn't started automatically
after reboot. It is also flaky and hard to debug because there are so
many activity state transitions in the log.

This CL moves lock task activity into a separate package and makes the
test query the state of that activity via a content provider.

The CL also moves it from DeviceAndProfileOwnerTest to DeviceOwnerTest
because there is little value in running this test in multiple
configuraitons - all the permission related checks are exercised by the
device side test in the new infra.

The CL aslo incorporates a check from another ignored test: that an
attempt to launch settings doesn't break out of lock task mode.

Bug: 354777547
Flag: EXEMPT test fix
Test: atest CtsDevicePolicyManagerTestCases:com.android.cts.devicepolicy.DeviceOwnerTest#testLockTaskAfterReboot
Change-Id: Icc7445929c3e3627946d0ac3ef9c3efc5f735658