[RESTRICT AUTOMERGE] audio: CtsAudioRecordPermissionTests cleanup
- Add additional actions to control foregroundness.
- Rename ACTION_CLEANUP for consistency.
- Send silence event when recording starts silenced
- Drop beginning frames of recording to warmup, to avoid spurious
  silence
- Modify activity/service cleanup to handle multiple packages
- Modify test to assume silence state following triggering record
- Modify test to explicitly go foreground
- Dismiss keyguard on sleep
- Ensure that background app isn't frozen
- Migrate to AMS commands instead of shell to allow for sending
 parcels. Requires adding view all packages perm.
- Rename action for clarity plus logging modification
- Add extends STS TestCase class
- Rename for consistency with other CTS modules
- Correct platform_api/sdk_version conflict to correct test_api version
- Data deps correct API version
- No instant mode

For all of our record permission tests, verify that the op state of the
app correctly matches the expected silencing status from audioserver.
This is critical to ensure the microphone indicator is accurate wrt to
the actual data that the app is receiving.

Add tests which verify that audioserver correctly handles the various
initial proc/fgs/api combos when recording is started.

Correct test api access on user

AudioRecordPermission test remove appop cooldowns

Bug: 293603271
Bug: 325912429
Test: atest AudioRecordPermissionTests
Flag: EXEMPT security
Change-Id: Ie8aef3b761f24e8b643649144251e054fafcb84e