add telecom and telephony static flag libs to the BP
The provider CTS module was failing with a `NoClassDefFoundError`
due to missing dependencies for the
`com.android.server.telecom.flags.Flags` class and the
related junit library for the annotations.

This commit adds the following dependencies to the
`Android.bp` file:
- `telecom_flags_core_java_lib`
- `telephony_flags_core_java_lib`
- `flag-junit`

These dependencies provide the necessary classes and junit
library to resolve the runtime error and allow the test to
execute correctly.

Flag: EXEMPT test fix
Fixes: 420349694
Test: atest CtsProviderTestCases
Change-Id: If3d544fc079059a2fe3289e149fee058943bfb40