run the entire CtsTelecomCujTestCases module on every change
An incident occured where a test failure bubbled up after
an isolated test change in this module.  This should have
never happened because this test module should run every
time there is a change in it to ensure there are no
breaking changes.

To prevent this from happening in the future, this change
will make it so that the entire module is run whenever
there is a change to it.

Flag: EXEMPT adding non-flaggable file
Fixes: 420942341
Test: none
Change-Id: I27a0d053545b08dc51d48d267f3547e96ef48b73