CTS: Skip AVBToTv hostside tests for playback devices
Temporarily disable AVBToTv hostside tests for playback devices because we
disabled AVB when system audio device is a TV for playback devices.
In CTS hostside test, the DUT is either connected to a TV or a dummy head
that mimics TV. Unlike service tests, we cannot re-enable AVB for hostside test.

Test: atest HdmiCecAvbToTvTest
Bug: b/406050353, b/421350435
Flag: EXEMPT test only fix
Change-Id: If6ec00b9d13bbaa09169b67422e19f7af029fc21