[Flexiglass] Fix the broken CTS test by showing bouncer
Currently, `enterAndConfirmLockCredential()` unlocks a device by
entering a lock credential without bouncer shown. It is a
bug filed in b/417307529

This issue is exposed by test failure when the Flexiglass is on. With
the Flexiglass, pressing the enter key shows the bouncer instead of
confirming the credential.

This CL shows the bouncer through a shell command before entering
the credential.

It works with Flexiglass on when tested locally.

Bug: 412771434
Flag: EXEMPT bugfix
Change-Id: I07039e0f74d2a994777daaf0e4195d296e581b32