Wait on LAST_ACTIVE_USER_ID signal for user switch
Previously, CarUserManagerHostTest was waiting on the current user to be switched to signify the completion of a user switch. However, the user switch does not deterministically end at this point and caused failures in subsequent tests. Now, the test should wait for android.car.LAST_ACTIVE_USER_ID, which is set by CarUserService when a user switch completes. 

Increasing the wait timeout also ensures that there is adequate time to wait for LAST_ACTIVE_USER_ID to be set.

Test: atest CtsCarHostTestCases
Bug: 413157225
Flag: EXEMPT test fix 
Change-Id: Ib7a3a5808b547b8ad3c372489c3236229894f4f9