Finish switching to initial user before removing test users
BaseDevicePolicyTest#tearDown can cause test failures. Due to a
race condition, it may attempt to remove the test user before switching
back to the initial user.

The fix is to wait for the switch back to the initial user to complete.
This can be done by calling waitForBroadcastIdle(), as
suggested in the documentation for switchUser().

Bug: 418981701
Test: atest com.android.cts.devicepolicy.DeviceOwnerTest#testCreateAndManageUser_CannotStopCurrentUser
Flag: EXEMPT test fix only
Change-Id: I491921d76e82255bbab5d69e27ea17cce9657f44