Fix CTS-on-GSI exclusions for key attestation performance tests.
The exclusions were introduced in ag/33201491, but with a typo in
the target name for the performance tests.

Also, the entries in the `cts-on-gsi-include-apks.xml` file use
"exclude-filter" instead of "include-filter", probably due to
copy-pasting, so updating those as well. From my understanding,
that file doesn't affect the test exclusions but could affect GRF
testing, according to a bug from two years ago (b/284917528).

Bug: 410503137
Bug: 424062514
Flag: TEST_ONLY
Test: Treehugger
Change-Id: I805af650c058b69320a6a382e011cbd77a5a05c7