Relax one-sided ranging measurement variation
Currenlty the non-11mc ranging measurement variation is 4 meters. In
one-sided RTT measurements, it is difficult to set a consistent KPI
because the peer device's variable turnaround time introduces
uncertainty. This issue arises because only the Device Under Test (DUT)
is responsible for estimating both transmission and reception timing. So
increase the variation to 6 meters.

Bug: 419756721
Flag: TEST_ONLY
Test: atest android.net.wifi.rtt.cts.WifiRttTest
Change-Id: I90d878d7bdbce503315b4c61cf8d0a2dba218dc6