[cts] Remove split tests from platinum tests
Currently they cannot run on 16K page devices because of the misalignment of the libraries, and according to the VSR requirement, all the shared libraries should be 16k aligned, so eventually, we will have to compile all the libraries in this test with 16k alignment. Removing the test from platinum for now until we can recompile all the libraries.

BUG: 422878242
Test: presubmit
FLAG: TEST_ONLY
Change-Id: Ia5fe66440101c1574f69a081e910637f42480f8b