Post CarrierPrivilege callbacks from Binder
To avoid a case where the main thread could be blocked waiting on
carrier privilege callbacks, but the callback change is posted to the
main thread, make a speculative fix to post the carrier privileges
callbacks directly from the binder thread. This is lightweight work, and
it will make sure that anything waiting on the main thread for carrier
privileges (directly or indirectly) OR that's just doing heavy work on
the main thread preventing service before the timeout won't break the
test.

This may or may not help, but it definitely won't hurt™!

Note: tentatively validated on ag/34091113.

Bug: 425756925
Flag: EXEMPT bugfix
Test: atest CtsTelephonyTestCases
Change-Id: Id2007b473651ce7f2e57b77c02688ab154f129db