Make CredentialManagerRestoreSettingsHostSideTest compatible with HSUM
The test is currently hardcoded to run on user 0 and is therefore explicitly skipped on headless system user mode (HSUM) devices where this assumption is invalid.

This CL updates the test logic to be user-aware.
It now dynamically fetches the current user ID during setup and uses it for critical operations.

Also, SwitchUserTargetPreparer's target user is changed from system user to the first full user to be compatible with HSUM.
On non-HSUM devices, this change has no functional impact as the system user 0 is also the first full user.

Bug: 425872836
Bug: 426128237
Test: atest CredentialManagerRestoreSettingsHostSideTest
Flag: TEST_ONLY
Change-Id: Id13f2642ef20c2826fd0ddd5e646425fa01b737d