Remove check for no availability changed after listener removal.
This test is quite flaky; it looks like the synchronization model in the
ImsRcsManager is such that the binder thread which removes the listener
and the one that notifies of capability status changed are racing and
causing a 35% flake.  Removing the check for availability change after
the listener is unregistered to improve test stability.

Fixes: 425887618
Test: THIS is a test fix.
Flag: EXEMPT test fix.
Change-Id: I80095c824452693a149c1af49da8c33ef3556aec