fix(LoginActivityCommonTestCase.testFocusLostBeforeResponse): add wait to reduce flakiness.
The test testFocusLostBeforeResponse was flaky because it relies on clearing the focus of the field
before the Service responds to the Request.
However, the Service mock was just sending the response as soon as the request arrives, so depending
on the thread execution order, the response could arrive before or after the clearFocus operation, resulting
in a flaky test.

Now, wait for 10ms to reduce flakiness.

b/399047815

Test passed: https://android-build.corp.google.com/abtd/run/L97700030013192309/?referrer=email
Flag: EXEMPT CTS fix
Change-Id: Ib75aa855cab9687fec7b4add21075ce0625bb0d6