Fix race condition in HotwordDetectionServiceBasicTest
The AppOps listener in HotwordDetectionServiceBasicTest was suffering
from a race condition between the listener callback and the verification
method, which was reading the op string which was set after the latch
was tripped.

Remove the entire op string validation concept, as we only
listen for the OP_RECORD regardless, so that is the only op we would
ever get.

Also, refactor the listener to be a static inner class to share with
HotwordDetectionServiceStressTest which has the same issue.
Additionally, include a fail-fast check to prevent reuse of the same
detector instance for multiple await calls, which would also cause
races.

Bug: 373567948
Test: atest HotwordDetectionServiceBasicTest
Test: atest HotwordDetectionServiceStressTest
Flag: EXEMPT test only
Change-Id: I673f90af5eb332d1a62f3d37f93c19d3016296d6