veq CTS: push resources as per test needs
Before the start of all tests, all resources are pushed to the device.
This is done explicitly by the test and not done via TargetPreparer.
This is causing issues in shard mode. Modify the implementation such
that the resources are pushed by each test as per its needs.

Bug: 411304697
Test: atest CtsVideoEncodingQualityHostTest
Flag: TEST_ONLY
Change-Id: I22f4603eeeb3244af93d11c088a9426cce557ddf