Fix the POST_PROMOTED_NOTIFICATIONS CTS test
The permission was actually in the manifest of the wrong package. 🤦

I don't quite understand why the test was flaky instead of always-failing (although running locally it was like 8/10 failures). Might have something to do with the instrumentation? In any case it works now, and we can remove the superfluous sleeps.

Fixes: 423334874
Test: atest --rerun-until-failure 10 CtsNotificationTestCases:android.app.notification.current.cts.NotificationManagerTest#testCanPostPromotedNotifications_apiEnabled
Flag: TEST_ONLY
Change-Id: I105eac050e2e6500d5b8bb1fc6f352871089ea07