[ADI][cts] change test behavior after adding the policy delegate support
Access control for setting the policy should happen regardless whether there
is a verifier installed on the device. Updating the CL to match the
expected behavior of the API.

This CL also adds a simple test to check that the test cannot query the
policy delegate.

FLAG: android.content.pm.verification_service
BUG: 420966373
BUG: 399436145
Test: atest android.content.pm.cts.DeveloperVerifierServiceTest

Change-Id: I0be5b160e6b960d7d687b019a6bff25bb3637e89