Add tests for route provider changes while dialog is open
These test that the dialog updates appropriately when the output
switcher is open with a route selected, and the provider then either
releases the session or forces a transfer to a different route.

Also skip two stream expansion tests when the output switcher redesign
flag is turned off, which had been just failing.

Test: atest OutputSwitcherTest
Flag: TEST_ONLY
Bug: 399188388
Change-Id: I9c30d5930205c0aa0836e77b2af85a5da256e047

checkpoint

Change-Id: I16a05cfd9afe397aa967457efe49a789cd51c808