Disable SM log check for call log integration
Ensure that we do not try to verify logging for self-managed calls when
the integrated_call_logs flag is enabled. This behavior is being
deprecated and replaced by the transactional call logging being done as
part of this feature. This was causing some test failures in
SelfManagedConnectionServiceTest when the flag was enabled.

Bug: 428056641
Flag: com.android.server.telecom.flags.integrated_call_logs
Test: atest SelfManagedConnectionServiceTest
Change-Id: I2b490655acfae2401a51817615574850e91b6896