Adjust the timing in Database CTS tests
There are intermittent failures in testCloseIdleConnection() that can
be explained by a test that races with the IdleConnectionHandler job
in SQLiteConnectionPool.  The test opens a database, pauses, verifies
that the connection is open, and waits for the connection to close.
The intermittent failures occur in the "verifies is open" stage.

This change does not fix the race but it makes the failure less likely
by removing the pause before checking that the connection is
open.

Also, there are two tests that check the idle connection timeout.  The
tests had different timing parameters; this change gives them the same
parameters.  The timing parameters are taken from the test that has
not been reported as flaky.

Flag: EXEMPT bug-fix
Bug: 425552429
Test: atest
 * CtsDatabaseTestCases

Change-Id: I90b13e684fea84357fa71f22dbd8a1a174e73303