Revert^2 "WebView: add custom TargetPreparer"
No change to logic. This adds a custom TargetPreparer for WebView CTS
tests which runs during the test setup in order to fetch debugging info
about the WebView service. This runs 'adb shell dumpsys webviewupdate',
captures the output in a text file, and logs this file as a tradefed
artifact which will be visible on test infra.

The main goal of this change is to log the WebView version in a single
location which is easily discoverable.

The log file will follow a pattern like:
"log/invocation_*/inv_*/webview-version_*.txt".

This TestPreparer is integrated into CtsWebkitTestCases,
CtsSdkSandboxWebkitTestCases, and CtsHostsideWebViewTests.

This reverts commit ba598ca04c990198c03d07fae6c5fd2f6c2db449 (reland of
http://ag/33993942).

Reason for revert: fixed by adding 'host_common_data'

Fixes: 417818812
Test: https://android-build.corp.google.com/abtd/run/L70600030013669187
Test: atest CtsWebkitTestCases
Test: atest CtsSdkSandboxWebkitTestCases
Test: atest CtsHostsideWebViewTests
Flag: TEST_ONLY
Change-Id: Iccdef72f0205b1f9798e4bc4bbdca53305087d9d