Remove references to flag android.sdk.major_minor_versioning_scheme
The flag android.sdk.major_minor_versioning_scheme was finalized as part
of Android 36.0. This version of Android has now been pushed to AOSP, so
the flag can be safely removed.

Note: because the flag is exported, the flag *declaration* must remain
(for now).

Bug: 430194198
Test: treehugger
Test: atest android.content.pm.parsing.cts.host.UsesSdkTest CtsOsTestCases:BuildTest
Flag: EXEMPT removal of finalized flag
Change-Id: I7523bea76ea51b4fb3fd6c32eeb6209fcc727b89