ITS: Get zoomMethod used by the default camera app
Added function to check the zoomMethod used by the default camera app.
If the value of android.control.zoomMethod is AUTO, it then checks if
the android.scaler.cropRegion is set or not. If it is set the function
returns the zoom_method as 'cropRegion' and 'zoomRatio' otherwise.
The next cl will have the change to transform the cropRegion to
zoomRatio and use that if the zoomMethod equals 'cropRegion'.

Bug: 407826153
Test: Tested locally
FLAG: TEST_ONLY
Change-Id: Id378cf164ca2eee07c0e9fb55bbacecd0e040aa0