Adding CTS test coverage for propagation of property to Telecom.
This property exists in the Connection already, we're not propagating it
to the Call.Details as well.  Note that I didn't flag this in the test
since the propery just gets passed through regardless if the public API
definition is available or not.  It is not really possible to flag that
mapping.

Test: Added CTS coverage for new property.
Flag: com.android.server.telecom.flags.remotely_hosted_property
Bug: 430293079
Change-Id: I2947b5c268551c197035dab4456cd9c45419df49