WebView: add keyevent HOME to dismiss menu popup
We noticed that if a device is already unlocked, then pressing
`keyevent MENU` will cause the device to display a user-visible menu
dialog. On some devices, this user-visible menu dialog can steal focus
and cause the tests to fail. This adds an extra `keyevent HOME` to
dismiss the dialog so that our tests can hopefully run with proper
window focus.

Bug: 419073650
Bug: 413487496
Test: atest CtsWebkitTestCases CtsHostsideWebViewTests CtsSdkSandboxWebkitTestCases (on a phone)
Test: This was also tested on another form factor in b/419073650#comment16
Flag: TEST_ONLY
Change-Id: I64563b8ce1b0dde7345faa447735d05891f7689d