Go to lockscreen before launching the translucent activity.
Currently, we launch this activity, then go to lockscreen, which we do by turning the screen off and on again. There's a bug in both legacy and flexiglass code (not sure what) that results in the wallpaper window not being visible if you turn the screen off and on again with a transparent, occluding activity visible. This happens both from unlocked and from lockscreen, if a transparent activity is at the top of the stack.

Legacy code usually aborts the screen off quickly enough to avoid the bug. Flexiglass does not, hence the 100% failure w/ flexi and the flakiness in legacy code. Not sure if this is is a SysUI, WM, or wallpaper engine bug.

The test is not actually testing this case though and is just testing whether transparent activities show the wallpaper, so I'm updating the test to reflect that.

Fixes: 416718983
Test: atest KeyguardTests
Flag: com.android.systemui.scene_container
Change-Id: I50715f3caef275b523bc05ceb205fe6b7213e08b