Revert^2 "Continue IME test migration from hostside"
This reverts commit ecf77733e9212ae293808cb3de6230ced23b44e6.

Reason for revert: Root cause of b/407125743 fixed through
  I3ddcc49947f68a18216551bc9434ce453debdeea

This continues the migration of the IME tests from hostside to
deviceside, to increase reliability.

testInputUnbindsOnImeStopped is moved to
InputMethodServiceTest#testImeUnbindsOnImeStopped. The previous
implementation was uninstalling the IME, however this IME is still
required for subsequent tests. Instead, the new test will simply force
stop the IME, which is equivalent for the purposes of this tet stash post, as
mentioned in the original bug (73142831) and original CL [1].

testInputUnbindsOnAppStop already matched most of the scenario from
runImeVisibilityTestWhenForceStopPackage, with the assertions on
bind/unbind added now.

testImeSwitchingWithoutWindowFocusAfterDisplayOffOn is moved to
ImeSwitchingTest#testImeSwitchingWithoutWindowFocusAfterDisplayOffOn,
now using LockScreenSession to reliably control the lock screen
transition.

  [1]: Ice2a94b9f6209de36dd340cdd2b9415efafb8618

Flag: EXEMPT testfix
Bug: 323251870
Test: atest ImeSwitchingTest#testImeUnboundAfterUninstalling
  ImeSwitchingTest#testImeSwitchingWithoutWindowFocusAfterDisplayOffOn
  KeyboardVisibilityControlTest#testImeInvisibleWhenForceStopPkgProcess_Full
 KeyboardVisibilityControlTest#testImeInvisibleWhenForceStopPkgProcess_Instant

Change-Id: I8432fb9c3876e1295aab2753aefaa70c94e5e86f