Check IME Navigation Bar visibility at test start
This adds a check that the IME navigation bar is visible at the start of
testOnCustomImeSwitcherButtonRequestedVisible_gestureNav. There is no
strong enforcement on the IME Navigation Bar being present and visible.
In [1] a code path that allowed directly disabling it was removed,
however it is still possible that other implementations might hide or
remove it entirely.

As the test relies on the IME Navigation Bar being visible from the
start, this adds an assumption on this state, to silently skip the test,
instead of considering this a compatibility issue.

Flag: EXEMPT testfix
Bug: 424078428
Test: atest InputMethodServiceTest#testOnCustomImeSwitcherButtonRequestedVisible_gestureNav
Change-Id: I4625801de84a2c36e0d6956511cfbdad6185a46d