CTS Verifier: unprocessed input for tap to tone
Some OEMs add noise cancellation to certain input streams.
When we send a noise in our tap to tone tests for edge detection,
this sound may get filtered out.

This CL changes the input preset to unprocessed so noise cancellation
is no longer a factor for this test.

Bug: 430535193
Bug: 429934040
Test: Run Tap to Tone test
Flag: TEST_ONLY
Change-Id: I329262214e7b31588993b0711e3054784c7f353f