Fix OutputSwitcherTest flakiness
The appCallsTransferToWhileDialogOpen_dialogUpdates test was
exhibiting some flakiness, and it seems to be due to the fact that we
were immediately asserting that it had a state description of
"Connected" as soon as we found the node in the UI, and sometimes we'd
get unlucky and it hadn't been set yet. This change fixes that by
adding a polling check with timeout before failing the test. It also
includes a small refectoring to better share some code.

Bug: 431070333
Test: atest OutputSwitcherTest
Flag: TEST_ONLY
Change-Id: I2e0a0957e2993e08578969eafdec447fed9dcabf