Fix MediaCasTest
Remove redundant close session calls in the resource challenge tests.
The sessions get auto closed with the Cas close calls.

The resource challenge tests update the cas info to have a maximum of 1
resource, but do not set it back to the original Integer.MAX_VALUE,
which could cause a fail in the testClearKeyApis() which requires two
sessions.

Bug: 381864924
Test: atest android.media.misc.cts.MediaCasTest
Flag: EXEMPT bugfix
Change-Id: Id0cc35bba2cdd72477f11c667f4edba6d413b2dd